move generate_reports to reports module
